chore: changed directory structure
