+++ src/renderer/features/operation/sign/ui/Signing/Signing.tsx
@@ -1,17 +1,17 @@
+import React from 'react';
+
 import { SigningProps } from '../../model';
 import { SigningType } from '@renderer/domain/shared-kernel';
 import { VaultSigning } from '../VaultSigning/VaultSigning';
 
+export const SigningFlow: Record<SigningType, (props: SigningProps) => React.JSX.Element | null> = {
+  [SigningType.MULTISIG]: (props) => <VaultSigning {...props} />,
+  [SigningType.PARITY_SIGNER]: (props) => <VaultSigning {...props} />,
+  [SigningType.WATCH_ONLY]: (props) => null,
+};
+
 export const Signing = (props: SigningProps) => {
   const signingType = props.accounts[0].signingType;
 
-  const getSigningFlow = () => {
-    if (signingType === SigningType.MULTISIG || SigningType.PARITY_SIGNER) {
-      return <VaultSigning {...props} />;
-    }
-
-    return <></>;
-  };
-
-  return getSigningFlow();
+  return SigningFlow[signingType](props);
 };
